+++ trunk/src/Resources/Core/RsrcAsyncLoadingReqsHandler.h
@@ -33,8 +33,8 @@ class RsrcAsyncLoadingReqsHandler
 		/// @param maxTime The max time to spend serving finished requests. If for example there are many that need
 		/// more time than the max the method will return. The pending requests will be served when it will be called
 		/// again.
-		/// In ms
-		void postProcessFinishedRequests(uint maxTime);
+		/// In seconds
+		void postProcessFinishedRequests(float maxTime);
 	
 	private:
 		AsyncLoader al; ///< Asynchronous loader
